SNOW-791036 * fix package names
