--- src/main/java/com/snowflake/core/SnowflakeLogger.java
@@ -1,6 +1,0 @@
-package com.snowflake.core;
-
-/** Allow plugins to pass their Maven/Gradle logger to core package using this interface */
-public interface SnowflakeLogger {
-  void info(String s);
-}
