--- src/main/java/com/snowflake/snowflake_gradle_plugin/extensions/FunctionContainer.java
@@ -1,15 +1,0 @@
-package com.snowflake.snowflake_gradle_plugin.extensions;
-
-/**
- * Container class for users to specify their functions for the CREATE FUNCTION DDL executed by the
- * plugin
- */
-public class FunctionContainer extends UserDefinedContainer {
-  public FunctionContainer(String name) {
-    super(name);
-  }
-  // Returns a concrete data class from user input to the container
-  public FunctionConcrete concrete() {
-    return new FunctionConcrete(getName(), getArgs(), getHandler(), getReturns());
-  }
-}
