--- src/main/java/com/snowflake/snowflake_gradle_plugin/extensions/UserDefinedContainer.java
@@ -1,20 +1,0 @@
-package com.snowflake.snowflake_gradle_plugin.extensions;
-
-import lombok.Getter;
-import lombok.Setter;
-
-/** Base class users to configure Function and Procedure definitions */
-@Setter
-@Getter
-public abstract class UserDefinedContainer {
-  public UserDefinedContainer(String name) {
-    this.name = name;
-  }
-
-  private String name;
-  private String handler;
-  private String returns;
-  private String[] args;
-
-  public abstract  UserDefinedConcrete concrete();
-}
