--- src/test/java/com/snowflake/core/LoggerMock.java
@@ -1,6 +1,0 @@
-package com.snowflake.core;
-
-/** Represents the logger passed from Maven/Gradle plugins */
-public class LoggerMock {
-  public void info(String s) {}
-}
