+++ src/test/java/com/snowflake/plugins/udf/core/LoggerMock.java
@@ -1,0 +1,6 @@
+package com.snowflake.plugins.udf.core;
+
+/** Represents the logger passed from Maven/Gradle plugins */
+public class LoggerMock {
+  public void info(String s) {}
+}
